requirements.txt
flask==3.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pypdf==4.2.0
python-docx==1.1.2
requests==2.32.3
numpy==1.26.4
